Add files via upload
The dataset contains prices and attributes for approximately 54,000 round-cut diamonds .
There are 53,940 diamonds in dataset with 10 features. I develop a predictive model to estimate diamond prices based on various features such as carat , cut, clarity, color, and depth. The result of Diamond Price Prediction is a predictive model that can forecast diamond prices based on various characteristics of the diamonds. The best score of the testing model would depend on the specific techniques and algorithms used in the analysis. However, a higher score would indicate that the model is more accurate in predicting diamond prices, which would lead to more informed pricing and inventory management decisions and increased revenue and profitability for the company.